build(vercel): rejig to accommodate build norms
- Shard `app` from `main`, to allow others to obtain a Nest instance
  - Also add export of response handler wrapper for vercel
- Add `api/index.js` shim for vercel, `public/robots.txt` to trick vercel
- Rewrite Vercel config to make use of shim
