+++ api/index.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const { vercelHandler } = require('../dist/app')
+
+module.exports = vercelHandler
