+++ src/main.ts
@@ -1,29 +1,7 @@
-import 'dotenv/config'
-import { NestFactory } from '@nestjs/core'
-import { getBotToken } from 'nestjs-telegraf'
-import cookieParser from 'cookie-parser'
-
-import { AppModule } from './app.module'
-import { ConfigService } from '@nestjs/config'
-import { DatabaseService } from './database/database.service'
-import { session } from 'telegraf'
-import { executeMigration } from './database/utils/pg.utils'
+import { appPromise } from './app'
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule)
-  const configService = app.get<ConfigService>(ConfigService)
-  const databaseService = app.get<DatabaseService>(DatabaseService)
-
-  app.use(cookieParser())
-
-  const bot = app.get(getBotToken())
-  bot.use(
-    session({
-      store: databaseService.store,
-      getSessionKey: (ctx) => `${ctx.from.id}`,
-    }),
-  )
-  app.use(bot.webhookCallback(configService.get<string>('bot.path')))
-  await app.listen(3000)
+  const app = await appPromise
+  app.listen(3000)
 }
 bootstrap()
